requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
plotly>=5.15.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
geopy>=2.4.0
pydantic>=2.0.0
